[JKM] Dagger 오류 수정
