--- app/google-services.json
@@ -1,39 +1,0 @@
-{
-  "project_info": {
-    "project_number": "1087171431521",
-    "project_id": "untakengoodsdb",
-    "storage_bucket": "untakengoodsdb.appspot.com"
-  },
-  "client": [
-    {
-      "client_info": {
-        "mobilesdk_app_id": "1:1087171431521:android:e8be26eee91e6ea2ae6e20",
-        "android_client_info": {
-          "package_name": "com.emart24"
-        }
-      },
-      "oauth_client": [
-        {
-          "client_id": "1087171431521-q256depo6f3s25498sffn6gq7an7hks9.apps.googleusercontent.com",
-          "client_type": 3
-        }
-      ],
-      "api_key": [
-        {
-          "current_key": "AIzaSyDmb5H7gsuGIOEwYBit8oRmwfa5KSpPObQ"
-        }
-      ],
-      "services": {
-        "appinvite_service": {
-          "other_platform_oauth_client": [
-            {
-              "client_id": "1087171431521-q256depo6f3s25498sffn6gq7an7hks9.apps.googleusercontent.com",
-              "client_type": 3
-            }
-          ]
-        }
-      }
-    }
-  ],
-  "configuration_version": "1"
-}
